requirements.txt
numpy
torch>=1.12.0
torchvision>=0.13.0
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX
streamlit
tqdm
scipy
